+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Add support for expiring the password for both the "root" and "stick" user.
 - Add offline ClamAV database to image build.
 - Add support for firewalld.
+- Add support for disabling CPU mitigations in the Linux kernel.
 
 ## 2.1.0